requirements.txt
weaviate-client==4.9.3
numpy==1.22.0
scikit-learn==1.5.0
